apps/pdf-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
reportlab==4.2.5
python-dotenv==1.0.1
pydantic==2.10.3
